requirements.txt
Django==3.1
psycopg2-binary
Pillow
djangorestframework
python-decouple
djangorestframework-simplejwt
drf-yasg